pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "collatz"
version = "0.1.0"
dependencies = [
    "pytest==8.0.2",
    "pygraphviz==1.12",
    "pillow==10.2.0",
]

[project.scripts]
collatz = "collatz.cli:run"
